jgrapht-core/src/test/java/org/jgrapht/GraphsTest.java     -New unit test are moved from the top to the bottem, so the old ones doesn't look like removed     -DummyGraph and DummyGraphType moved indo separated files
jgrapht-core/src/test/java/org/jgrapht/DummyGraph.java
and
jgrapht-core/src/test/java/org/jgrapht/DummyGraphType.java
    -These classes are used to test in the same time neither directed nor undirected graphs (an exception is expected in that case)

jgrapht-core/src/test/java/org/jgrapht/GraphTestsTest.java
    -Some new unit test for class jgrapht-core/src/main/java/org/jgrapht/GraphTests.java

jgrapht-core/src/main/java/org/jgrapht/GraphTests.java
    -Check whether the graph is directed in the following functions:
        -isWeaklyConnected
        -isStronglyConnected

        In the case of an undirected graph will an IllegalArgumentException thrown, just like in the other functions, because these functions are just for directed graphs defined.
